backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
langchain-classic>=0.0.1
langchain-community>=0.3.0
langchain-chroma>=0.1.4
langchain-huggingface>=0.1.0
langchain-google-genai>=2.0.0
langchain-text-splitters>=0.3.0
rank-bm25>=0.2.2
pypdf>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0
